+++ src/utils/DataBuilder.java
@@ -7,7 +7,7 @@ import entities.UMovie;
 import java.time.LocalDate;
 
 public class DataBuilder {
-    public static void AddMovie(int id, int collectionId, String collectionName, int budget, String language, String title, LocalDate releaseDate, int revenue){
+    public static void AddMovie(int id, int collectionId, String collectionName, long budget, String language, String title, LocalDate releaseDate, long revenue){
         Pelicula p = new Pelicula(id,budget, null, language, title, releaseDate, revenue, collectionId, collectionName);//va si generos
         UMovie.insertarPeliculas(p);
